moved playqueryexecutor and its util out of players.py
